backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform
google-cloud-firestore
google-adk